@@ -20,7 +20,7 @@
         "optimize-autoloader": true
     },
     "require": {
-        "php": ">=7.4"
+        "php": ">=8.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.4",
